Each class should be its own file, even the main one!
